[API] changes for Julia bindings (#1016)
This PR brings some changes to the API, in order to facilitate the creation of Julia bindings for SIRIUS in a future SIRIUS.jl package.

Apart from the additions of new functions to sirius_api.cpp (necessary for the DFTK integration), the main addition consists of a header file for the API: sirius_c_headers.h. It turns out that having a header file drastically simplifies the process of creating Julia bindings for a C/C++ library.

As things stand, this header file is generated by a script (generate_c_headers.py), in the same spirit as the Fortran API. As a result, it is not used for compilation, but only as source material for the automatic generation of Julia bindings.